buildable project and testpypi testing
