+++ .gitignore
@@ -1,4 +1,7 @@
-__pycache__
+*__pycache__
+*.egg-info
+build
+dist
 
 flake.nix
 flake.lock
